fix: address PR #119 follow-up review
